requirements.txt
fastapi
uvicorn[standard]
requests
redis[hiredis]
python-dotenv